Refactoring: moved all the rotation matrix functions to a separate file
